Fixed data size bug

+++ appserver/src/app.cpp
@@ -401,14 +401,14 @@ void App::printException(const std::exception &e) const
 
 bool App::sendAck(std::shared_ptr<zmq::socket_t> socket)
 {
-    int ack = 1;
+    uint8_t ack = 1;
     zmq::message_t msg(&ack, sizeof(uint8_t));
     return socket->send(msg);
 }
 
 bool App::recvAck(std::shared_ptr<zmq::socket_t> socket)
 {
-    int ack = 0;
+    uint8_t ack = 0;
     size_t receivedSize = socket->recv(&ack, sizeof(uint8_t));
     if (receivedSize <= 0 || ack != 1) {
         return false;
